Move Helper class into bob.util pkg
