+++ src/main/java/bob/parser/Parser.java
@@ -2,7 +2,6 @@ package bob.parser;
 import java.io.IOException;
 import java.time.LocalDateTime;
 
-import bob.Helper;
 import bob.command.Command;
 import bob.command.DeadlineCommand;
 import bob.command.DeleteCommand;
@@ -13,6 +12,7 @@ import bob.command.MarkCommand;
 import bob.command.TodoCommand;
 import bob.command.UnmarkCommand;
 import bob.command.WrongCommandException;
+import bob.util.Helper;
 
 public class Parser {
     public static Command parse(String userInput) throws WrongCommandException, IOException {
