+++ src/main/java/bob/task/Event.java
@@ -1,7 +1,7 @@
 package bob.task;
 import java.time.LocalDateTime;
 
-import bob.Helper;
+import bob.util.Helper;
 
 public class Event extends Task {
     private LocalDateTime from;
