+++ src/main/java/bob/util/Helper.java
@@ -1,10 +1,8 @@
-package bob;
+package bob.util;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 public class Helper {
-
-
     public static String datetimeToReadable(LocalDateTime dateTime) {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM d yyyy h:mma");
         return dateTime.format(formatter);
